+++ internal/uci/utils.go
@@ -1,0 +1,8 @@
+package uci
+
+func abs(n int) int {
+	if n < 0 {
+		return -n
+	}
+	return n
+}